Silvio/leut 248/statemachine

+++ code_SemperKI/services/service_AdditiveManufacturing/handlers/model.py
@@ -9,6 +9,7 @@ Contains: handling of model files
 import logging
 from datetime import datetime
 
+from django.conf import settings
 from django.http import JsonResponse, HttpResponse
 from django.views.decorators.http import require_http_methods
 from django.utils import timezone
@@ -23,6 +24,7 @@ from code_SemperKI.definitions import ProcessDescription, ProcessUpdates, DataTy
 from code_SemperKI.handlers.projectAndProcessManagement import updateProcessFunction, getProcessAndProjectFromSession
 from code_SemperKI.connections.content.postgresql import pgProcesses
 from code_SemperKI.handlers.files import deleteFile
+from code_SemperKI.connections.content.manageContent import ManageContent
 
 from ..definitions import ServiceDetails
 from ..connections.postgresql import pgService
@@ -65,6 +67,7 @@ def uploadModel(request):
         model[fileID][FileObjectContent.URI] = ""
         model[fileID][FileObjectContent.createdBy] = userName
         model[fileID][FileObjectContent.path] = filePath
+        model[fileID][FileObjectContent.remote] = False
 
         returnVal = s3.manageLocalS3.uploadFile(filePath, request.FILES.getlist(fileName)[0])
         if returnVal is not True:
@@ -88,12 +91,14 @@ def uploadModel(request):
 
 #######################################################
 @require_http_methods(["DELETE"])
-def deleteModel(request, processID):
+def deleteModel(request, projectID, processID):
     """
     Delete the model and the file with it, if not done already
     
     :param request: request
     :type request: HTTP POST
+    :param projectID: The ID of the project
+    :type projectiD: str
     :param processID: process ID
     :type processID: Str
     :return: Successful or not
@@ -101,29 +106,23 @@ def deleteModel(request, processID):
 
     """
     try:
-        modelOfThisProcess = {}
-        modelExistsAsFile = False
-        currentProjectID, currentProcess = getProcessAndProjectFromSession(request.session,processID)
-        if currentProcess != None:
-            modelOfThisProcess = currentProcess[ProcessDescription.serviceDetails][ServiceDetails.model]
-            if modelOfThisProcess[FileObjectContent.id] in currentProcess[ProcessDescription.files]:
-                modelExistsAsFile = True
-        else:
-            if manualCheckifLoggedIn(request.session) and manualCheckIfRightsAreSufficient(request.session, deleteModel.__name__):
-                currentProcess = pgProcesses.ProcessManagementBase.getProcessObj(processID)
-                modelOfThisProcess = currentProcess.serviceDetails[ServiceDetails.model]
-                if modelOfThisProcess[FileObjectContent.id] in currentProcess.files:
-                    modelExistsAsFile = True              
-            else:
-                return HttpResponse("Not logged in or rights insufficient!", status=401)
-
+        # TODO maybe rewrite to support multiple model files
+        contentManager = ManageContent(request.session)
+        interface = contentManager.getCorrectInterface(deleteModel.__name__)
+        if interface == None:
+            return HttpResponse("Not logged in or rights insufficient!", status=401)
+        currentProcess = interface.getProcessObj(projectID, processID)
+        modelOfThisProcess = currentProcess.serviceDetails[ServiceDetails.model]
+        if modelOfThisProcess[FileObjectContent.id] in currentProcess[ProcessDescription.files]:
+            modelExistsAsFile = True
+            
         if modelExistsAsFile:
             deleteFile(request, processID, modelOfThisProcess[FileObjectContent.id])
         
         changes = {"changes": {}, "deletions": {ProcessUpdates.serviceDetails: {ServiceDetails.model: modelOfThisProcess[FileObjectContent.id]}}}
-        message, flag = updateProcessFunction(request, changes, currentProjectID, [processID])
+        message, flag = updateProcessFunction(request, changes, projectID, [processID])
         if flag is False:
-            return JsonResponse({}, status=401)
+            return HttpResponse("Insufficient rights!", status=401)
         if isinstance(message, Exception):
             raise message
 
@@ -132,3 +131,47 @@ def deleteModel(request, processID):
     except (Exception) as error:
         loggerError.error(f"Error while deleting file: {str(error)}")
         return HttpResponse("Failed", status=500)
+
+#######################################################
+@require_http_methods(["GET"])
+def getModelRepository(request):
+    """
+    Get previews and file names of 3D models from out curated repository.
+
+    :param request: GET request
+    :type request: HTTP GET
+    :return: JSON with names of files, link to preview files and link to files itself
+    :rtype: JSONResponse
+    
+    """
+    content = s3.manageRemoteS3Buckets.getContentOfBucket("ModelRepository")
+    outDict = {}
+    for elem in content:
+        path = elem["Key"]
+        splitPath = path.split("/")[1:]
+        if len(splitPath) > 1:
+            if "license" in elem["Metadata"]:
+                license = elem["Metadata"]["license"]
+            else:
+                license = ""
+            if splitPath[0] in outDict:
+                if "Preview" in splitPath[1]:
+                    outDict[splitPath[0]]["preview"] = s3.manageRemoteS3Buckets.getDownloadLinkPrefix()+elem["Key"].replace(" ", "%20")
+                else:
+                    outDict[splitPath[0]]["file"] = s3.manageRemoteS3Buckets.getDownloadLinkPrefix()+elem["Key"].replace(" ", "%20")
+                if license != "" and outDict[splitPath[0]]["license"] == "":
+                    outDict[splitPath[0]]["license"] = license
+            else:
+                outDict[splitPath[0]] = {"name": splitPath[0], "license": "", "preview": "", "file": ""}
+                if "Preview" in splitPath[1]:
+                    outDict[splitPath[0]]["preview"] = s3.manageRemoteS3Buckets.getDownloadLinkPrefix()+elem["Key"].replace(" ", "%20")
+                else:
+                    outDict[splitPath[0]]["file"] = s3.manageRemoteS3Buckets.getDownloadLinkPrefix()+elem["Key"].replace(" ", "%20")
+                if license != "" and outDict[splitPath[0]]["license"] == "":
+                    outDict[splitPath[0]]["license"] = license
+                
+
+    return JsonResponse(outDict)
+
+
+
